+++ src/main/java/com/example/AgenciaTurismo/repository/IHotelReservaRepository.java
@@ -1,0 +1,8 @@
+package com.example.AgenciaTurismo.repository;
+
+import com.example.AgenciaTurismo.model.ReservarHotel;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IHotelReservaRepository extends JpaRepository<ReservarHotel, Long> {
+
+}
