+++ src/main/java/com/example/AgenciaTurismo/service/HotelService.java
@@ -25,7 +25,7 @@ public class HotelService implements IHotelService {
     @Autowired
     IHotelRepository repository;
 
-    private final List<HotelReservedDTO> hotelReserve = new ArrayList<>();
+
     @Override
     public List<HotelDTO> listarHotels() {
         List<HotelDTO> listHotel = repository.findAll().stream()
@@ -53,55 +53,6 @@ public class HotelService implements IHotelService {
         hotelAvailable.setAvailableHotelDTO(availableHotel);
 
         return hotelAvailable;
-    }
-
-    @Override
-    public TotalHotelReservationDTO reserved(FinalHotelReservationDTO finalHotelReservationDTO) {
-
-        if (this.reserveSaved(finalHotelReservationDTO)) {
-            throw new IllegalArgumentException("La reserva ya está realizada.");
-        }
-
-        this.roomCapacity(finalHotelReservationDTO.getHotelReservationDTO());
-
-
-        HotelConsultDTO hotelBuscado = new HotelConsultDTO(finalHotelReservationDTO.getHotelReservationDTO().getDateFrom(),
-                finalHotelReservationDTO.getHotelReservationDTO().getDateTo(),
-                finalHotelReservationDTO.getHotelReservationDTO().getDestination());
-
-        List<HotelDTO> availableHotel = this.validarHotelesDisponibles(hotelBuscado);
-
-
-        HotelDTO hotelToReserved = null;
-        for (HotelDTO hotel : availableHotel) {
-            if (hotel.getDestination().equalsIgnoreCase(finalHotelReservationDTO.getHotelReservationDTO().getDestination())
-                    && hotel.getDateFrom().equals(finalHotelReservationDTO.getHotelReservationDTO().getDateFrom())
-                    && hotel.getDateTo().equals(finalHotelReservationDTO.getHotelReservationDTO().getDateTo())) {
-                hotelToReserved = hotel;
-                break;
-            }
-        }
-        if (hotelToReserved == null) {
-            throw new IllegalArgumentException("No se encontró ningún hotel que coincida con los criterios de reserva.");
-        }
-
-        Double amount = (hotelToReserved.getPriceForNight() * finalHotelReservationDTO.getHotelReservationDTO().getPeopleAmount());
-
-        Double interest = this.calcInterest(amount, finalHotelReservationDTO.getHotelReservationDTO().getPaymentMethodDTO().getDues(),
-                finalHotelReservationDTO.getHotelReservationDTO().getPaymentMethodDTO().getType());
-
-        Double total = amount + interest;
-
-        TotalHotelReservationDTO totalHotelReservationDTO = new TotalHotelReservationDTO();
-        totalHotelReservationDTO.setAmount(amount);
-        totalHotelReservationDTO.setInterest(interest);
-        totalHotelReservationDTO.setTotal(total);
-        totalHotelReservationDTO.setFinalHotelReservation(finalHotelReservationDTO);
-        totalHotelReservationDTO.setStatusCode(new StatusCodeDTO(201, "El proceso terminó satisfactoriamente"));
-
-        hotelReserve.add(new HotelReservedDTO(totalHotelReservationDTO));
-
-        return totalHotelReservationDTO;
     }
 
 
@@ -153,79 +104,7 @@ public class HotelService implements IHotelService {
         return new ResponseDTO("Hotel eliminado con éxito");
     }
 
-            //METODOS PARA REUTILIZAR
-    @Override
-    public Boolean reserveSaved(FinalHotelReservationDTO finalHotelReservationDTO) {
-        for (HotelReservedDTO reservaGuardada : hotelReserve) {
-            FinalHotelReservationDTO reservaExistente = reservaGuardada.getHotelReserved().getFinalHotelReservation();
-            if (reservaExistente.equals(finalHotelReservationDTO)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public List<HotelReservedDTO> hotelSaved() {
-        return hotelReserve;
-    }
-
-    //Evalua el metodo de pago ingresado para poder hacer el carlculo de la reserva.
-    @Override
-    public Double calcInterest(Double amount, Integer dues, String type) {
-
-        if (type.equalsIgnoreCase("Debit") || type.equalsIgnoreCase("Credit")) {
-            if (type.equalsIgnoreCase("Debit") && dues > 1) {
-                throw new IllegalArgumentException("No puede pagar en cuotas con tarjeta de debito.");
-            } else
-                switch (dues) {
-                    case 1:
-                        return 0.0;
-                    case 2, 3:
-                        return amount * 0.05;
-                    case 4, 5, 6:
-                        return amount * 0.10;
-                    case 7, 8, 9, 10, 11, 12:
-                        return amount * 0.20;
-                    default:
-                        throw new IllegalArgumentException("Número de cuotas no válido.");
-                }
-        } else {
-            throw new IllegalArgumentException("Tipo de pago no válido.");
-        }
-
-    }
-
-    //comparar tIpo de habitacion con cantidad de personas ingresadas
-    public Boolean roomCapacity(HotelReservationDTO reservation) {
-        Double people;
-
-
-        switch (reservation.getRoomType()) {
-            case "Single":
-                people = 1D;
-                break;
-            case "Doble":
-                people = 2D;
-                break;
-            case "Triple":
-                people = 3D;
-                break;
-            case "Múltiple":
-                people = 4D;
-                break;
-            default:
-                people = 0D;
-                break;
-        }
-        if (!people.equals(reservation.getPeopleAmount())) {
-            throw new IllegalArgumentException("La cantidad de personas no coincide con el tipo de habitación.");
-        } else if (people != reservation.getPeopleDTO().size()) {
-            throw new IllegalArgumentException("El tipo de habitación seleccionada no coincide con la cantidad de personas que se alojarán en ella.");
-        }
-        return true;
-    }
-
+    //METODOS PARA REUTILIZAR
     // Validando Hoteles disponible
     @Override
     public List<HotelDTO> validarHotelesDisponibles(HotelConsultDTO hotelConsultDTO) {
