+++ src/main/java/com/example/AgenciaTurismo/service/IHotelService.java
@@ -1,13 +1,9 @@
 package com.example.AgenciaTurismo.service;
 
 import com.example.AgenciaTurismo.dto.HotelDTO;
-import com.example.AgenciaTurismo.dto.HotelReservationDTO;
-import com.example.AgenciaTurismo.dto.HotelReservedDTO;
-import com.example.AgenciaTurismo.dto.request.FinalHotelReservationDTO;
 import com.example.AgenciaTurismo.dto.request.HotelConsultDTO;
 import com.example.AgenciaTurismo.dto.response.HotelAvailableDTO;
 import com.example.AgenciaTurismo.dto.response.ResponseDTO;
-import com.example.AgenciaTurismo.dto.response.TotalHotelReservationDTO;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -18,7 +14,6 @@ public interface IHotelService {
 
     HotelAvailableDTO hotelesDisponibles (HotelConsultDTO hotelConsultDTO);
 
-    TotalHotelReservationDTO reserved(FinalHotelReservationDTO finalHotelReservationDTO);
 
     // CREATE
     ResponseDTO createHotel(HotelDTO hotelDTO);
@@ -27,15 +22,6 @@ public interface IHotelService {
     ResponseDTO updateHotel(Long id, HotelDTO hotelDTO);
     //DELETE
     ResponseDTO deleteHotel (Long id);
-
-    //METODOS PARA VALIDAR
-    Boolean reserveSaved(FinalHotelReservationDTO finalHotelReservationDTO);
-
-    List<HotelReservedDTO> hotelSaved();
-
-    Double calcInterest(Double amount, Integer dues, String type);
-
-    Boolean roomCapacity(HotelReservationDTO reservation);
 
     List<HotelDTO> validarHotelesDisponibles(HotelConsultDTO hotelConsultDTO);
