+++ src/main/resources/application.properties
@@ -11,7 +11,7 @@ spring.jpa.hibernate.ddl-auto=create-drop
 # CONNECTION
 # Database's User & Password
 spring.datasource.username=root
-spring.datasource.password=123456
+spring.datasource.password=
 
 # URL & Name
 spring.datasource.url=jdbc:mysql://localhost:3306/agenciaturismo_db
